requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
cron-descriptor==1.4.3
cryptography==42.0.5
Django==4.2.10
django-admin==2.0.2
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-excel-response2==3.0.6
django-ninja==0.22.2
django-ninja-extra==0.19.8
django-ninja-jwt==5.2.5
django-redis==5.4.0
django-role-permissions==3.2.0
django-six==1.0.5
django-smtp-ssl==1.0
django-timezone-field==6.1.0
excel-base==1.0.4
gunicorn==21.2.0
idna==3.6
injector==0.21.0
isoweek==1.3.3
kombu==5.3.6
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pycparser==2.21
pydantic==1.10.14
PyJWT==2.8.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
redis==5.0.3
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing-extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
xlwt==1.3.0
python-decouple==3.8
django-rest-framework==0.1.0